move FailingJSONEncoder to separate file, formatting
